+++ src/config.js
@@ -94,7 +94,9 @@ class Config {
   }
 
   getEventConfig(type, subType) {
-    return this.#eventMap[this.generateKey(type, subType)];
+    return this.#eventMap[this.generateKey(type, subType)]
+      ? { ...this.#eventMap[this.generateKey(type, subType)] }
+      : undefined;
   }
 
   isCapOutboxEvent(type) {
@@ -770,6 +772,10 @@ class Config {
 
   get isMultiTenancy() {
     return !!cds.requires.multitenancy;
+  }
+
+  get _rawEventMap() {
+    return this.#eventMap;
   }
 
   /**