+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Updated cron-parser to major version ^5.0.0
 
+### Fixed
+
+- fixed keep alive for some edge cases
+
 ## v1.9.1 - 2025-02-25
 
 ### Fixed