+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - fixed keep alive for some edge cases
+- fix after deployments - parallel checking of changes to periodic events and processing of open events
 
 ## v1.9.1 - 2025-02-25