+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## v1.9.2 - 2025-03-XX
+## v1.9.3 - 2025-03-18
+
+### Fixed
+
+- removed not wanted info log statements
+
+## v1.9.2 - 2025-03-18
 
 ### Added